requirements.txt
requests
yt-dlp
whisper
pytest
sqlalchemy
apscheduler
python-dotenv
